requirements.txt
hikari-lightbulb
async-covid